refactore folder structure for maven
