+++ GraphQL-MusicStore-Maven/src/main/java/com/example/graphqlmusicstoremaven/GraphQlMusicStoreMavenApplication.java
@@ -3,7 +3,10 @@ package com.example.graphqlmusicstoremaven;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(
+        scanBasePackageClasses = {
+                com.example.graphqlmusicstoremaven.graphql.ApplicationSpecificSpringComponentScanMarker.class
+        })
 public class GraphQlMusicStoreMavenApplication {
 
     public static void main(String[] args) {
