+++ GraphQL-MusicStore-Maven/src/main/java/com/example/graphqlmusicstoremaven/graphql/ApplicationSpecificSpringComponentScanMarker.java
@@ -1,0 +1,4 @@
+package com.example.graphqlmusicstoremaven.graphql;
+
+public interface ApplicationSpecificSpringComponentScanMarker {
+}
